Added tests for the SparseOneHotEncoder
